Fix the nothing provides kernel-devicetree build error
Issue: LIN1018-4213

We have to make sure that the PREFERRED_PROVIDER_virtual/kernel is set
for all the distro types. Otherwise we would get the following build
errors:
  ERROR: Nothing RPROVIDES 'kernel-devicetree'
  NOTE: Runtime target 'kernel-devicetree' is unbuildable, removing...
  Missing or unbuildable dependency chain was: ['kernel-devicetree']
  ERROR: Required build target 'core-image-sato' has no buildable providers.
  Missing or unbuildable dependency chain was: ['core-image-sato', 'kernel-devicetree']

This is because the linux-dummy is used due to that the PREFERRED_PROVIDER_virtual/kernel
is not set in the distro conf such as nodisto and poky.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>