Change the KMACHINE to marvell-cn96xx-sdkv5.4
Issue: LINCD-8613

So far the LTS21 (v5.10/standard/cn-sdkv4.18/cn96xx) and CD
(v5.10/standard/cn-sdkv5.4/octeon) use the same machine definition in
the kernel-cache. In order to fix a bug in LTS21, we have factored out
the kernel options which are specific to the CD kernel to a separate
file and introduce a new machine definition to include these kernel
options, so we should set the KMACHINE to that in the BSP layer.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>