marvell-cn9xxx: remove setting WRL_RECIPES from templates
Issue: LINUXEXEC-26075

Some customers want access to use recipes that are included in
WRL_RECIPES but they do not use templates. By moving the updates to
WRL_RECIPES outside of the default template we can simplify this
usecase.

We move the changes to WRL_RECIPES of recipes found in this layer to
the layer.conf. Since this is a Wind River layer we can host the
WRL_RECIPES here to save having to edit multiple layers when we add or
remove recipes to this layer.

Signed-off-by: Sakib Sajal <sakib.sajal@windriver.com>